Refactor: Create resume
