+++ src/components/profile/ProfileContainer.tsx
@@ -9,12 +9,16 @@ import { APP_CONSTANTS } from "@/lib/constants";
 import Loading from "../Loading";
 import ResumeTable from "./ResumeTable";
 import { toast } from "../ui/use-toast";
+import { PlusCircle } from "lucide-react";
+import { Button } from "../ui/button";
 
 function ActivitiesContainer() {
   const queryParams = useSearchParams();
   const recordsPerPage = APP_CONSTANTS.RECORDS_PER_PAGE;
 
   const [resumes, setResumes] = useState<Resume[]>([]);
+  const [resumeDialogOpen, setResumeDialogOpen] = useState(false);
+
   const [resumeToEdit, setResumeToEdit] = useState<Resume | null>(null);
   const [totalResumes, setTotalResumes] = useState(0);
   const [currentPage, setCurrentPage] = useState(
@@ -55,13 +59,15 @@ function ActivitiesContainer() {
     loadResumes(currentPage);
   }, [currentPage, loadResumes]);
 
+  const createResume = () => {
+    setResumeToEdit(null);
+    setResumeDialogOpen(true);
+  };
+
   const onEditResume = (resume: Resume) => {
     const _resumeToEdit = { id: resume.id, title: resume.title };
     setResumeToEdit(_resumeToEdit);
-  };
-
-  const resetResumeToEdit = () => {
-    setResumeToEdit(null);
+    setResumeDialogOpen(true);
   };
 
   return (
@@ -69,10 +75,22 @@ function ActivitiesContainer() {
       <CardHeader className="flex-row justify-between items-center">
         <CardTitle>Profile</CardTitle>
         <div className="flex items-center">
+          <Button
+            size="sm"
+            variant="outline"
+            className="h-8 gap-1"
+            onClick={createResume}
+          >
+            <PlusCircle className="h-3.5 w-3.5" />
+            <span className="sr-only sm:not-sr-only sm:whitespace-nowrap">
+              Create Resume
+            </span>
+          </Button>
           <CreateResume
+            resumeDialogOpen={resumeDialogOpen}
+            setResumeDialogOpen={setResumeDialogOpen}
             reloadResumes={reloadResumes}
             resumeToEdit={resumeToEdit}
-            resetResumeToEdit={resetResumeToEdit}
           />
         </div>
       </CardHeader>
